Fix values with commas breaking CSV output.
If a value is found to have a comma in it, the value will get
wrapped in quotation marks, and any quotation marks in the value
will be escaped.

Set it up to possilbly allow users to specify the encapsulation
character in the future.  There wasn't any documentation on the
supported opts (and no validation), so I wasn't sure how this
was supposed to be implemented.

I'd like to move the file to use all space indention or tab-based
indention, but I figured if we wanted to do that, it could be in
another commit.